refactor: extract components, hook, styles
