+++ src/types.ts
@@ -1,0 +1,7 @@
+interface IFormInputs {
+  email: string;
+  password: string;
+  remember: boolean;
+}
+
+export type { IFormInputs };
